+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added new `--bundle` flag to the `operator-sdk scorecard` command to support bundle validation testing using the validation API (https://github.com/operator-framework/api). ([#1916](https://github.com/operator-framework/operator-sdk/pull/1916)
 - Added new `log` field to the `operator-sdk scorecard` v1alpha2 output to support tests that produce logging. ([#1916](https://github.com/operator-framework/operator-sdk/pull/1916)
 - Added new `bundle validation` test to the `operator-sdk scorecard` OLM tests. ([#1916](https://github.com/operator-framework/operator-sdk/pull/1916)
+- Added new `CSV validation` test to the `operator-sdk scorecard` OLM tests.  This test uses a new operator-framework/api validation API to perform a more comprehensive test on the CSV. ([#1916](https://github.com/operator-framework/operator-sdk/pull/1916)
 
 ### Changed
 - Upgrade minimal Ansible version in the init projects from `2.4` to `2.6`. ([#2107](https://github.com/operator-framework/operator-sdk/pull/2107))
@@ -20,6 +21,7 @@
 ### Deprecated
 
 ### Removed
+- Removed `CRs Have At Least 1 Example` scorecard OLM test.  This test was replaced by a new `CSV validation` test. ([#1916](https://github.com/operator-framework/operator-sdk/pull/1916)
 
 ### Bug Fixes
 - Fix issue faced in the Ansible based operators when `jmespath` queries are used because it was not installed. ([#2252](https://github.com/operator-framework/operator-sdk/pull/2252))